LayoutTests/imported/mathml-in-html5/mathml/relations/html5-tree/namespace-1.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Namespaces</title>
<link rel="help" href="http://www.mathml-association.org/MathMLinHTML5/S2.html#SS1.SSS1">
<meta name="assert" content="Verify whether some elements are in the right namespaces.">
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>
  var namespaces = {
    "XHTML": "http://www.w3.org/1999/xhtml",
    "SVG": "http://www.w3.org/2000/svg",
    "MathML": "http://www.w3.org/1998/Math/MathML"
  };
  for (var c in namespaces) {
    var test_namespace = async_test("Test elements in " + c + " namespace");
    window.addEventListener("DOMContentLoaded", function() {
      var el = document.getElementsByClassName(c);
      for (var i = 0; i < el.length; i++) {
        test_namespace.step(function() {
          assert_equals(el[i].namespaceURI, namespaces[c]);
        });
      }
    });
    test_namespace.done();
  }
</script>
</head>
<body>
  <math class="MathML" xmlns="http://www.w3.org/1998/Math/MathML"></math>
  <math class="MathML"></math>
  <math><mtext class="MathML"></mtext></math>
  <svg width="0" height="0">
    <foreignObject height="0" width="0">
      <math class="MathML"></math>
    </foreignObject>
  </svg>
  <math>
    <semantics>
      <mtext><span class="XHTML"></span></mtext>
      <annotation-xml encoding="SVG1.1">
        <svg class="SVG" width="0" height="0"></svg>
      </annotation-xml>
      <annotation-xml encoding="image/svg+xml">
        <svg class="SVG" width="0" height="0"></svg>
      </annotation-xml>
      <annotation-xml encoding="application/xhtml+xml">
        <span class="XHTML"></span>
      </annotation-xml>
      <annotation-xml encoding="text/html">
        <span class="XHTML"></span>
      </annotation-xml>
    </semantics>
  </math>
</body>
</html>
